+++ NEWS.md
@@ -1,8 +1,8 @@
 # tidycmprsk (development version)
 
-* Now returning the the number at risk, cumulative number of events, and the cumulative number of censored observations in `tidy.cuminc()`.
+* Now returning the the number at risk, cumulative number of events, and the cumulative number of censored observations in `tidy.tidycuminc()`.
 
-* Updated the default value to `tidy.cuminc(conf.int = TRUE)` from `FALSE`
+* Updated the default value to `tidy.tidycuminc(conf.int = TRUE)` from `FALSE`
 
 * Update the `crr()` print method for more clarity.